Public Documentation

Documentation for AcousticWaveCPML.jl's public interface.

Public Interface

AcousticWaveCPML.Receivers — Type

Type representing a multi-receiver configuration for a wave propagation shot.

AcousticWaveCPML.Receivers — Method

Receivers(positions::Matrix{<:Real})

Create a single shot wave propagation receivers configuration from receivers positions.

AcousticWaveCPML.Sources — Type

Type representing a multi-source configuration for a wave propagation shot.

AcousticWaveCPML.Sources — Method

Sources(
    positions::Matrix{<:Real},
    t0s::Vector{<:Real},
    srctfs::Vector{<:Function},
    freqdomain::Real
)

Create a single shot wave propagation source configuration from source positions, time-functions and a frequency domain.

AcousticWaveCPML.rickersource1D — Method

rickersource1D(t::Real, t0::Real, f0::Real)

Ricker source time function for current time t, activation timet0and dominating frequencyf0`.

AcousticWaveCPML.Acoustic1D.solve1D — Method

solve1D(
    lx::Real,
    lt::Real,
    vel::Vector{<:Real},
    srcs::Sources,
    recs::Receivers;
    halo::Integer = 20,
    rcoef::Real = 0.0001,
    ppw::Real = 10.0,
    do_bench::Bool = false,
    do_vis::Bool = false,
    nvis::Integer = 5,
    gif_name::String = "acoustic1D",
    plims::Vector{<:Real} = [-1.0, 1.0]
)

Solve 1D acoustic wave equation on a model with length lx [m], for final time lt [sec] with the velocity model vel [m/s]. The size of vel implicitly specifies the number of grid points to use for the simulation.

Sources and receivers positions are specified by srcs and recs.

Return the final time pressure field as a vector and populate the receivers seismograms with the recorded traces.

Arguments

halo::Integer = 20: the number of CPML layers.

rcoef::Real = 0.0001: the reflection coeffiecient for CPML layers.

ppw::Real = 10.0: minimum number of points per wavelength admissible.

do_bench::Bool = false: if benchmark instead of computation.

do_vis::Bool = false: enables plotting.

nvis::Integer = 5: plotting time step frequency.

gif_name::String = "acoustic1D": name of animation plot.

plims::Vector{<:Real} = [-1.0, 1.0]: limits for pressure values in plot.

AcousticWaveCPML.Acoustic2D.solve2D — Method

TODO

AcousticWaveCPML.Acoustic2D_Threads.solve2D — Method

TODO

AcousticWaveCPML.Acoustic2D_CUDA.solve2D — Method

TODO

AcousticWaveCPML.Acoustic2Dmulti_Threads.solve2D_multi — Method

TODO

AcousticWaveCPML.Acoustic2Dmulti_CUDA.solve2D_multi — Method

TODO

AcousticWaveCPML.Acoustic3D.solve3D — Method

TODO

AcousticWaveCPML.Acoustic3D_Threads.solve3D — Method

TODO

AcousticWaveCPML.Acoustic3D_CUDA.solve3D — Method

TODO

AcousticWaveCPML.Acoustic3Dmulti_Threads.solve3D_multi — Method

TODO

AcousticWaveCPML.Acoustic3Dmulti_CUDA.solve3D_multi — Method

TODO