Public Documentation

Documentation for AcousticWaveCPML.jl's public interface.

Public Interface

AcousticWaveCPML.Receivers — Type

Type representing a multi-receiver configuration for a wave propagation shot.

AcousticWaveCPML.Receivers — Method

Receivers(positions::Matrix{<:Real})

Create a single shot wave propagation receivers configuration from receivers positions.

AcousticWaveCPML.Sources — Type

Type representing a multi-source configuration for a wave propagation shot.

AcousticWaveCPML.Sources — Method

Sources(
    positions::Matrix{<:Real},
    t0s::Vector{<:Real},
    srctfs::Vector{<:Function},
    freqdomain::Real
)

Create a single shot wave propagation source configuration from source positions, time-functions and a frequency domain.

AcousticWaveCPML.gaussource1D — Method

gaussource1D(t::Real, t0::Real, f0::Real)

First derivative of gaussian source function for current time t, activation timet0and dominating frequencyf0`.

AcousticWaveCPML.rickersource1D — Method

rickersource1D(t::Real, t0::Real, f0::Real)

Ricker source time function for current time t, activation timet0and dominating frequencyf0`.

AcousticWaveCPML.Acoustic1D.solve1D — Method

TODO

AcousticWaveCPML.Acoustic2D.solve2D — Method

TODO

AcousticWaveCPML.Acoustic2D_Threads.solve2D — Method

TODO

AcousticWaveCPML.Acoustic2D_CUDA.solve2D — Method

TODO

AcousticWaveCPML.Acoustic2Dmulti_Threads.solve2D_multi — Method

TODO

AcousticWaveCPML.Acoustic2Dmulti_CUDA.solve2D_multi — Method

TODO

AcousticWaveCPML.Acoustic3D.solve3D — Method

TODO

AcousticWaveCPML.Acoustic3D_Threads.solve3D — Method

TODO

AcousticWaveCPML.Acoustic3D_CUDA.solve3D — Method

TODO

AcousticWaveCPML.Acoustic3Dmulti_Threads.solve3D_multi — Method

TODO

AcousticWaveCPML.Acoustic3Dmulti_CUDA.solve3D_multi — Method

TODO